- Create a container file for userOrdersList - Refactor userOrdersList helper function
